requirements.txt
apache-airflow==2.9.3
apache-airflow-providers-mongo>=4.1.0
pymongo>=4.7.2
requests>=2.32.3
pydantic>=2.7.0
python-dotenv>=1.0.1
pytest>=8.3.2
mongomock>=4.1.2
